Move "colour.RGB_Colourspace" class to "colour.models.rgb" module.
