+++ colour/models/__init__.py
@@ -3,18 +3,18 @@
 
 from __future__ import absolute_import
 
-from .rgb_colourspace import RGB_Colourspace
+from .cie_xyy import (
+    XYZ_to_xyY,
+    xyY_to_XYZ,
+    xy_to_XYZ,
+    XYZ_to_xy)
+from .rgb import RGB_Colourspace
 from .derivation import (
     normalised_primary_matrix,
     RGB_luminance_equation,
     RGB_luminance)
 from .dataset import *  # noqa
 from . import dataset
-from .cie_xyy import (
-    XYZ_to_xyY,
-    xyY_to_XYZ,
-    xy_to_XYZ,
-    XYZ_to_xy)
 from .cie_lab import XYZ_to_Lab, Lab_to_XYZ, Lab_to_LCHab, LCHab_to_Lab
 from .cie_luv import (
     XYZ_to_Luv,
@@ -31,15 +31,15 @@ from .common import XYZ_to_sRGB, sRGB_to_XYZ
 from .aces_rgb_idt import spectral_to_aces_relative_exposure_values
 from .ipt import XYZ_to_IPT, IPT_to_XYZ, IPT_hue_angle
 
-__all__ = ['RGB_Colourspace']
+__all__ = ['XYZ_to_xyY',
+           'xyY_to_XYZ',
+           'xy_to_XYZ',
+           'XYZ_to_xy']
+__all__ += ['RGB_Colourspace']
 __all__ += ['normalised_primary_matrix',
             'RGB_luminance_equation',
             'RGB_luminance']
 __all__ += dataset.__all__
-__all__ += ['XYZ_to_xyY',
-            'xyY_to_XYZ',
-            'xy_to_XYZ',
-            'XYZ_to_xy']
 __all__ += ['XYZ_to_Lab', 'Lab_to_XYZ', 'Lab_to_LCHab', 'LCHab_to_Lab']
 __all__ += ['XYZ_to_Luv',
             'Luv_to_XYZ',
