+++ colour/models/dataset/russell_rgb.py
@@ -25,7 +25,8 @@ from __future__ import division, unicode_literals
 import numpy as np
 
 from colour.colorimetry.dataset import ILLUMINANTS
-from colour.models import RGB_Colourspace, normalised_primary_matrix
+from colour.models import normalised_primary_matrix
+from colour.models import RGB_Colourspace
 
 __author__ = 'Colour Developers'
 __copyright__ = 'Copyright (C) 2013 - 2014 - Colour Developers'
